# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API circleoflinks",
        "description": "API de red social de profecionales ",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/circleoflinks-app/public"
        }
    ],
    "paths": {
...
        "/api/v1/ofertalaboral": {
            "get": {
                "tags": [
                    "Oferta laboral"
                ],
                "summary": "Obtener todos los registros de una oferta laboral",
                "description": "Devuelve información detallada sobre una oferta laboral específica.",
                "operationId": "1dd89fa7920273eb7416d69f9c59c501",
                "responses": {
                    "200": {
                        "description": "Oferta laboral encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "se necesita Dev Junior Front"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "trabajo desarrollo"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": " se necesita con urgencia"
                                        },
                                        "salary": {
                                            "type": "number",
                                            "example": 2
                                        },
                                        "date_expire": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-11"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oferta laboral no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Oferta laboral no encontrada con ID: {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ofertalaboral/{id}": {
            "get": {
                "tags": [
                    "Oferta laboral"
                ],
                "summary": "Obtener detalles de una oferta laboral",
                "description": "Devuelve información detallada sobre una oferta laboral específica.",
                "operationId": "5a96fc807bcf13b4a64127229d3baf70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la oferta laboral",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Oferta laboral encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "se necesita Dev Junior Front"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "trabajo desarrollo"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": " se necesita con urgencia"
                                        },
                                        "salary": {
                                            "type": "number",
                                            "example": 2
                                        },
                                        "date_expire": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-11"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oferta laboral no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Oferta laboral no encontrada con ID: {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}